requirements.txt
# VC Workflow Automation - Render Deployment Dependencies

# Web Framework
Flask==2.3.3
gunicorn==21.2.0

# Google APIs
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.23.4

# Anthropic AI
anthropic==0.7.8

# Document Processing
python-docx==0.8.11

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Scheduling
schedule==1.2.0

# Data Models
pydantic==2.5.0
typing-extensions==4.8.0